+++ R/table_continuous_lm.R
@@ -963,7 +963,7 @@ table_continuous_lm <- function(
   effect_size_digits = 2,
   p_digits = 3,
   decimal_mark = ".",
-  align = c("decimal", "auto", "center", "right"),
+  align = c("decimal", "center", "right"),
   output = c(
     "default",
     "data.frame",