+++ R/table_regression.R
@@ -824,7 +824,7 @@ table_regression <- function(
   fit_digits = 2L,
   ic_digits = 1L,
   decimal_mark = ".",
-  align = c("decimal", "center", "right", "auto"),
+  align = c("decimal", "center", "right"),
   padding = 0L,
   labels = NULL,
   title = NULL,